# services/api/src/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Hill90 API
  description: Hill90 platform API — agent management and lifecycle operations.
  version: 0.1.0
servers:
  - url: /
    description: Relative (behind Traefik reverse proxy)
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ModelPolicy:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        allowed_models:
          type: array
          items:
            type: string
        max_requests_per_minute:
          type: integer
          nullable: true
          description: Rate limit (requests per minute). NULL means unlimited.
        max_tokens_per_day:
          type: integer
          nullable: true
          description: Daily token budget. NULL means unlimited.
        model_aliases:
          type: object
          additionalProperties:
            type: string
          nullable: true
          description: Alias map (alias name → real model name). Single-pass resolution.
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        updated_by:
          type: string
          nullable: true
        created_by:
          type: string
          nullable: true
          description: Owner (user sub). NULL for platform policies.
...
paths:
...
  /provider-connections/{id}/models:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    get:
      summary: List provider models
      description: >
        Returns available models from the provider API for a given connection.
        Each model includes auto-detected type and capabilities.
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Provider model list (may include error field on failure)
          content:
            application/json:
              schema:
                type: object
                properties:
                  models:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                          description: Provider-prefixed model ID (e.g., openai/gpt-4o)
                        display_name:
                          type: string
                        detected_type:
                          type: string
                          enum: [chat, embedding, audio, image, transcription]
                        capabilities:
                          type: array
                          items:
                            type: string
                  provider:
                    type: string
                  error:
                    type: string
                    description: Error message if model listing failed
        "401":
          description: Not authenticated
        "403":
          description: Requires user role
        "404":
          description: Connection not found
...
  /model-policies:
    get:
      summary: List model policies
      description: Returns model policies. Admins see all; users see own + platform.
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Policy list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ModelPolicy"
        "401":
          description: Not authenticated
        "403":
          description: Requires user role

    post:
      summary: Create model policy
      description: Creates a new model policy. Admin policies are platform-scoped; user policies are owner-scoped. model_aliases can only be set by admins.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - allowed_models
              properties:
                name:
                  type: string
                description:
                  type: string
                allowed_models:
                  type: array
                  items:
                    type: string
                max_requests_per_minute:
                  type: integer
                  nullable: true
                max_tokens_per_day:
                  type: integer
                  nullable: true
                model_aliases:
                  type: object
                  additionalProperties:
                    type: string
                  description: Alias map (admin-only). Each key maps to an allowed_models entry.
      responses:
        "201":
          description: Policy created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ModelPolicy"
        "400":
          description: Invalid input
        "401":
          description: Not authenticated
        "403":
          description: Requires user role
        "409":
          description: Duplicate policy name
  /model-policies/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid

    get:
      summary: Get model policy
      description: Returns a single model policy. Users can see own + platform policies.
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Policy detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ModelPolicy"
        "401":
          description: Not authenticated
        "403":
          description: Requires user role
        "404":
          description: Policy not found

    put:
      summary: Update model policy
      description: Updates a model policy. Users can update own policies; admins can update any. model_aliases can only be set by admins.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                description:
                  type: string
                allowed_models:
                  type: array
                  items:
                    type: string
                max_requests_per_minute:
                  type: integer
                  nullable: true
                max_tokens_per_day:
                  type: integer
                  nullable: true
                model_aliases:
                  type: object
                  additionalProperties:
                    type: string
                  description: Alias map (admin-only). Each key maps to an allowed_models entry.
      responses:
        "200":
          description: Policy updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ModelPolicy"
        "401":
          description: Not authenticated
        "403":
          description: Requires user role (model_aliases requires admin)
        "404":
          description: Policy not found
        "409":
          description: Duplicate policy name

    delete:
      summary: Delete model policy
      description: Deletes a model policy. Fails if agents are assigned. Users can delete own; admins can delete any.
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Policy deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  deleted:
                    type: boolean
        "401":
          description: Not authenticated
        "403":
          description: Requires user role
        "404":
          description: Policy not found
        "409":
          description: Agents are assigned to this policy
...